DipTraceImporter: Extract the GUI and the CSV parsing from the importer.
